third_party/blink/web_tests/platform/generic/fast/multicol/widows-and-orphans-expected.txt

Testing widows and orphans. Any green lines should be at the bottom of pages/columns, and any red lines should be at the top of pages/columns.

test1 - Normal breaking
test2 - Basic Orphan
test3 - Basic Widow
test4 - Orphans causing Widows
test5 - Widows blocked by Orphan rule
test6 - Ridiculous values

PASS: test1 Block 1 Line 1 is correct.

PASS: test1 Block 2 Line 5 is correct.

PASS: test1 Block 4 Line 2 is correct.

PASS: test2 Block 1 Line 1 is correct.

PASS: test2 Block 2 Line 1 is correct.

PASS: test3 Block 1 Line 1 is correct.

PASS: test3 Block 2 Line 5 is correct.

PASS: test4 Block 1 Line 1 is correct.

PASS: test4 Block 2 Line 1 is correct.

PASS: test4 Block 3 Line 3 is correct.

PASS: test5 Block 1 Line 1 is correct.

PASS: test5 Block 2 Line 3 is correct.

PASS: test6 Block 1 Line 1 is correct.

PASS: test6 Block 2 Line 1 is correct.

PASS: test6 Block 3 Line 1 is correct.
